+++ packages/pois/package.js
@@ -1,4 +1,4 @@
-var version = '1.6.7';
+var version = '1.6.8';
 
 Package.describe({
   version: version,