+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 [Downloadable releases](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
-- core K.Place.showLoc() and K.Place.showLoc() support callaback
+...
+
+## v1.6.8
+- core K.Place.showLoc() and K.Place.showGeometry() support callaback
 - plugin osm K.Osm.findByLoc() support node and way
-- plugin import reduce multigeometyr to single, new method K.Util.sanitize.importGeometry()
+- plugin import reduce MultiGeometry to single one using new method K.Util.sanitize.importGeometry()
 
 ## v1.6.7
 - core new settings K.settings.public.map.checkinGeometry autocheckin inside place geometry